All -- Upgrade toolchain and fix lint issues
Migrate to vite-plus and TypeScript 6, upgrade
eslint config to v5, pin typescript-eslint overrides,
and add explicit tsconfig path aliases. Fix int
overflow on BytesIn/BytesOut by using long, void
floating navigate() promises, and apply trailing
comma formatting from new lint rules.

+++ Haproxy.Editor.Api/Haproxy.Editor.Abstractions/Data/DashboardSnapshot.cs
@@ -100,9 +100,9 @@ public sealed record RuntimeBackendStatus
 
 	public int SessionRate { get; init; }
 
-	public int BytesIn { get; init; }
+	public long BytesIn { get; init; }
 
-	public int BytesOut { get; init; }
+	public long BytesOut { get; init; }
 
 	public int HealthyServers { get; init; }
